update folder locations for melter, remove media, update sequencer to use eyebrows.

+++ imagizer.py
@@ -23,39 +23,36 @@ class Imagizer:
         self.frame_rate = frame_rate
         self.faces = []
 
-    def init(self, img_path, img_name, out_path, compress=True):
+    def init(self, img_path, img_name, out_path, compress=True, qual=50):
         self.out_path = out_path
-        self.load_image(img_path, img_name, compress=compress)
+        self.clear_dir(out_path)
+        self.load_image(img_path, img_name, compress=compress, qual=qual)
         self.create_melters()
 
-    def load_image(self, img_path, img_name, compress=True):
+    def load_image(self, img_path, img_name, compress=True, qual=50):
         self.img_path = img_path
         self.img_name = img_name
         self.compress = compress
-        self.reload_image()
+        self.reload_image(qual=qual)
 
-    def reload_image(self):
+    def reload_image(self, qual=90):
         if self.compress:
-            self.img_name = self.compress_image(self.img_path, self.img_name)
+            self.img_name = self.compress_image(self.img_path, self.img_name, qual=qual)
         if self.img_path.endswith(self.img_name):
             self.original_image = cv2.imread(self.img_path)
+            self.image = cv2.imread(self.img_path)
         else:
             self.original_image = cv2.imread(self.img_path + self.img_name)
-        self.image = cv2.imread(self.img_path + self.img_name)
+            self.image = cv2.imread(self.img_path + self.img_name)
         if self.melt_master:
             return self.melt_master.image_reset(self.image)
 
     def detect_faces(self):
         try:
-            self.faces = self.face_cascade.detectMultiScale(
-                self.image,
-                scaleFactor=1.1,
-                minNeighbors=5,
-                minSize=(30, 30)
-                # flags = cv2.CV_HAAR_SCALE_IMAGE
-            )
+            self.faces = self.face_cascade.detectMultiScale(self.image,scaleFactor=1.1,minNeighbors=5,minSize=(30, 30))
+            # flags = cv2.CV_HAAR_SCALE_IMAGE
 
-            print("{0} :: Found {1} faces!".format(self.img_name, len(self.faces)))
+            # print("{0} :: Found {1} faces!".format(self.img_name, len(self.faces)))
             return self.faces
         except Exception as e:
             print('woops detect_faces: %s' % self.img_path + self.img_name)
@@ -78,7 +75,6 @@ class Imagizer:
             self.melt_master.create_melter(face)
 
     def face_melt(self):
-        print('applying melt')
         max_height = 0
         for face in self.faces:
             if face[3] > max_height:
@@ -89,11 +85,9 @@ class Imagizer:
         return self.melt_master.melt_method('melt', 1)
 
     def pause(self, num_frames=10):
-        print('applying pause')
         return self.melt_master.melt_method('pause', 100, num_frames=num_frames)
 
     def face_static(self, frame_rate, override=False, full_frame=False, reverse=False, num_frames=130):
-        print('applying static')
         self.melt_master.set_frame_rate(frame_rate, self.inverse_frame_rate)
         if override:
             return self.melt_master.melt_method('static', 1, target_coarse=num_frames, img_override=self.original_image, full_frame=full_frame, reverse=reverse)
@@ -101,12 +95,10 @@ class Imagizer:
             return self.melt_master.melt_method('static', 1, target_coarse=num_frames, full_frame=full_frame, reverse=reverse)
 
     def face_flash(self):
-        print('applying face color glitch')
         self.melt_master.set_frame_rate(1, self.inverse_frame_rate)
         return self.melt_master.melt_method('flash', 1)
 
     def full_flash(self):
-        print('applying full color glitch')
         self.melt_master.set_frame_rate(1, self.inverse_frame_rate)
         return self.melt_master.melt_method('flash', 1, full_frame=True)
 
@@ -124,7 +116,6 @@ class Imagizer:
 
     def add_text(self):
         org = (int(self.image.shape[1]/6), int(8 * self.image.shape[0]/10))
-        print(org)
         top_left = org
         bottom_right = (org[0] + 220, org[1] + 20)
         img_name = self.melt_master.get_last_frame()
@@ -141,11 +132,9 @@ class Imagizer:
         # cv2.addText(self.original_image, "the technocratic", org, cv2.FONT_HERSHEY_DUPLEX)
 
     def full_loop(self):
-        print(os.curdir)
         for file_name in os.listdir(os.curdir + '/pics'):
             try:
                 if file_name.split('.')[1] in VALID_EXT:
-                    print(file_name)
                     self.face_melt()
             except Exception as e:
                 print(e)
@@ -158,22 +147,20 @@ class Imagizer:
 
     @staticmethod
     def clear_dir(dirname):
-        shutil.rmtree(dirname)
+        if os.path.exists(dirname):
+            shutil.rmtree(dirname)
         os.mkdir(dirname)
 
     @staticmethod
     def compress_image(file_path, file_name, qual=50, resize=True):
-
         if file_path.endswith(file_name):
-            foo = Image.open(file_path)
-        else:
-            foo = Image.open(file_path + file_name)
+            file_path = file_path.split(file_name)[0]
+        foo = Image.open(file_path + file_name)
         new_file_name = 'comp_' + str(qual) + file_name
         if resize:
             divisor = 1
             while foo.size[0] / divisor > 300 and foo.size[1] / divisor > 300:
                 divisor += 1
-            print('divisor %s' % divisor)
             foo = foo.resize((int(foo.size[0] / divisor), int(foo.size[1] / divisor)))
         # foo.thumbnail((foo.size[0] / 4, foo.size[1] / 4), Image.ANTIALIAS)
         foo.save(file_path + new_file_name, optimize=True, quality=qual)
